requirements viejo moderno.txt
pyTelegramBotAPI==4.12.0
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.4
pytz==2023.3
python-dotenv==1.0.0
lxml==4.9.3
html5lib==1.1